Move metrics management to another class
